Print number of deployments to screen
Since some services have multiple deployments, they will each have their
diff values written to a directory directly beneath `build/`, such as
`build/DEPLOYMENT_NAME/deployment`.

This is done rather than doing it underneath the service name for all of
the its services, which wouldn't be as helpful when determining the
total number of deployments.

This is helpful for knowing how many deployments to diff for a
particular service.